requirements.txt
streamlit
openai
anthropic
google-generativeai
pypdf
python-dotenv 
typing-extensions